requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
pylint==2.6.0
python-decouple==3.4
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
